moved ObjectPrintingTests in test project
