+++ fluent-api.sln
@@ -13,6 +13,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FluentMapping.Tests", "Samp
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Spectacle", "Samples\Spectacle\Spectacle.csproj", "{EFA9335C-411B-4597-B0B6-5438D1AE04C3}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{E4AEC989-D540-44CF-8896-23C7DC04CB4C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ObjectPrintingTests", "tests\ObjectPrintingTests\ObjectPrintingTests.csproj", "{71311FE6-B67B-4E3C-A076-3FF03A745629}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -35,6 +39,10 @@ Global
 		{EFA9335C-411B-4597-B0B6-5438D1AE04C3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EFA9335C-411B-4597-B0B6-5438D1AE04C3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EFA9335C-411B-4597-B0B6-5438D1AE04C3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{71311FE6-B67B-4E3C-A076-3FF03A745629}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{71311FE6-B67B-4E3C-A076-3FF03A745629}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{71311FE6-B67B-4E3C-A076-3FF03A745629}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{71311FE6-B67B-4E3C-A076-3FF03A745629}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -43,5 +51,6 @@ Global
 		{FEEA5AFE-459A-4D13-81D0-252E1A2E6F4E} = {6D308E4A-CEC7-4536-9B87-81CD337A87AD}
 		{8A7BB3EA-3E6A-4D04-A801-D5CD1620DA0D} = {6D308E4A-CEC7-4536-9B87-81CD337A87AD}
 		{EFA9335C-411B-4597-B0B6-5438D1AE04C3} = {6D308E4A-CEC7-4536-9B87-81CD337A87AD}
+		{71311FE6-B67B-4E3C-A076-3FF03A745629} = {E4AEC989-D540-44CF-8896-23C7DC04CB4C}
 	EndGlobalSection
 EndGlobal
