+++ tests/ObjectPrintingTests/ObjectPrinterTests.cs
@@ -1,10 +1,8 @@
-using System;
-using System.Collections.Generic;
 using System.Globalization;
 using FluentAssertions;
-using NUnit.Framework;
+using ObjectPrinting;
 
-namespace ObjectPrinting.Tests;
+namespace ObjectPrintingTests;
 
 [TestFixture]
 public class ObjectPrinterTests
