+++ tests/ObjectPrintingTests/PersonWithParent.cs
@@ -1,6 +1,4 @@
-using System;
-
-namespace ObjectPrinting.Tests;
+namespace ObjectPrintingTests;
 
 public class PersonWithParent
 {
