+++ tests/ObjectPrintingTests/PersonWithWeight.cs
@@ -1,6 +1,4 @@
-using System;
-
-namespace ObjectPrinting.Tests;
+namespace ObjectPrintingTests;
 
 public class PersonWithWeight
 {
